+++ examples/plot_permutation_test_for_classification.py
@@ -22,7 +22,6 @@ import pylab as pl
 from sklearn.svm import SVC
 from sklearn.cross_validation import StratifiedKFold, permutation_test_score
 from sklearn import datasets
-from sklearn.metrics import accuracy_score
 
 
 ##############################################################################
@@ -43,7 +42,7 @@ svm = SVC(kernel='linear')
 cv = StratifiedKFold(y, 2)
 
 score, permutation_scores, pvalue = permutation_test_score(
-    svm, X, y, accuracy_score, cv=cv, n_permutations=100, n_jobs=1)
+    svm, X, y, scoring="accuracy", cv=cv, n_permutations=100, n_jobs=1)
 
 print("Classification score %s (pvalue : %s)" % (score, pvalue))